+++ CHANGELOG.md
@@ -5,11 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2025-11-01
+
+### Added
+- Added line numbers to stack trace output
+
+### Changed
+- Tidied up code
+
 ## [1.1.0] - 2025-11-01
+
+### Added
+- Added colours to debug methods, and added optional highlighting to Trace method
 
 ### Changed
 - Changed assembly definition naming
-- Added colours to debug methods, and added optional highlighting to Trace method
 
 ## [1.0.0] - 2025-11-01